re-arranged code
